fix: windows missing iconf / MagicWand.dll
Couldn't get my hands on these old version dlls from trusted sources.
qrscan is MIT licensed and works on both sides.
