--- img/empty.txt
@@ -1,1 +1,0 @@
-this folder has to remain empty
